impl: initial idea for a safe garbage collector
We want to support arbitrary types, but currently, only the concrete
type `Object` is supported. For arbitrary type, we want to somehow keep
a list of `Alloc<'heap, dyn Traceable<'heap>>`, which we currently don't
know how to do. Rust disallows casting to types having a trait object as
a generic argument.